CategoryListView created
