+++ Aspire Budgeting/Views/CategoryDetailsView.swift
@@ -12,10 +12,6 @@ struct CategoryDetailsView: View {
   private let bannerHeight: CGFloat = 100
   private let cornerRadius: CGFloat = 24
 
-  private func getAuxillaryText(spent: AspireNumber, budgeted: AspireNumber) -> String {
-    "\(spent.stringValue) spent • \(budgeted.stringValue) budgeted"
-  }
-
     var body: some View {
       VStack {
         banner
@@ -67,39 +63,11 @@ extension CategoryDetailsView {
           .padding([.top, .horizontal])
 
         Text("Budget per category")
-          .font(.karlaBold(size: 15))
+          .font(.karlaBold(size: 18))
           .foregroundColor(.secondaryTextColor)
 
-        ScrollView {
-          ForEach(cardDetails.categories, id: \.self) { category in
-            VStack(alignment: .leading) {
-              HStack {
-                Text(category.categoryName)
-                  .font(.karlaBold(size: 14))
-                  .foregroundColor(.primaryTextColor)
-                Spacer()
-                AspireNumberView(number: category.available)
-              }
-              HStack {
-                Text(getAuxillaryText(spent: category.spent,
-                                      budgeted: category.budgeted))
-                  .font(.karlaRegular(size: 12))
-                  .foregroundColor(.secondaryTextColor)
-                Spacer()
-
-                Text("available")
-                  .font(.karlaRegular(size: 12))
-                  .foregroundColor(.secondaryTextColor)
-              }
-
-              AspireProgressBar(barType: .minimal,
-                                shadowColor: .gray,
-                                tintColor: cardDetails.tintColor,
-                                progressFactor: category.available /| category.monthly)
-            }
-            .padding([.bottom, .horizontal])
-          }
-        }
+        CategoryListView(categories: cardDetails.categories,
+                         tintColor: cardDetails.tintColor)
       }
 
     }
